[6] rename EmptyResponse->NoContentParser, remove parseJSON from DataParser, remove JSONArrayParser
